add confirmation dialog on file delete

+++ src/DZarsky.TS3Viewer2.Web/src/pages/admin/Files.tsx
@@ -16,9 +16,13 @@ export const FilesPage = () => {
             return
         }
 
-        await FileService.deleteApiV1Files(file.fullName).then(() => {
-            getSongList()
-        })
+        const confirm = window.confirm(`Are you sure you want to delete ${file.fullName}?`)
+
+        if (confirm) {
+            await FileService.deleteApiV1Files(file.fullName).then(() => {
+                getSongList()
+            })
+        }
     }
 
     async function getSongList() {
